Move translation layer to react-native-electron-modules
