+++ apps/menu-bar/electron/modules/mainRegistry.ts
@@ -1,4 +1,5 @@
+import { Registry } from 'react-native-electron-modules';
+
 import MenuBarModule from '../../modules/menu-bar/electron/main';
-import { Registry } from './RegistryTypes';
 
 export const MainModules: Registry = [MenuBarModule];
