+++ apps/menu-bar/electron/modules/preloadRegistry.ts
@@ -1,4 +1,4 @@
-import { Registry } from './RegistryTypes';
+import { Registry } from 'react-native-electron-modules';
 import MenuBarModule from '../../modules/menu-bar/electron/preload';
 
 export const PreloadModules: Registry = [MenuBarModule];
