+++ apps/menu-bar/electron/package.json
@@ -31,6 +31,7 @@
     "typescript": "~4.9.4"
   },
   "dependencies": {
-    "electron-squirrel-startup": "^1.0.0"
+    "electron-squirrel-startup": "^1.0.0",
+    "react-native-electron-modules":"1.0.0"
   }
 }
